# Fix py3 socket slots