@carsonip 2020-12-20
Or maybe we can once and for all ditch the instance dict, add timeout and hub to the mixin slots and add empty slots to socket2 and socket3.